bot/requirements.txt
wppconnect-python==1.0.3
requests==2.28.2
python-dotenv==1.0.0
flask==2.2.3
gunicorn==20.1.0